requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.30.6
gunicorn==21.2.0
pydantic==2.11.4
pydantic-settings==2.10.1
python-dotenv==1.1.0
openai==1.99.1
aiofiles==23.2.1
azure-storage-blob
SQLAlchemy==2.0.41
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
pyodbc==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4